작업사항(21/11/03~04)
1. Core/DX11_Graphics_Pipeline
StructuredBuffer 클래스 구현

2. Component
ParticleSystem 클래스 구현 중
ParticleUpdateShader 클래스 구현 중

3. Resource
Texture 클래스 수정
1) BindPipelineRW(UAV 전용) 구현

4. ShaderFiles
ParticleStruct.fx 추가
ParticleUpdate.fx 추가

5. TODO
캐릭터 스크립트 깔끔하게 구현
애니메이션 및 애니메이터 최적화
파티클 시스템 구현
표준/구조화버퍼 공부

+++ Project/DX11_2D_GameEngine_Lib/Mesh.cpp
@@ -139,7 +139,7 @@ void Mesh::CreateCircleMesh()
 {
 	//Vertex
 	//mesh 사이즈(x와 y 값) 중 작은 사이즈를 반지름으로 결정
-	float radius = m_width <= m_height ? m_width : m_height;
+	float radius = m_width <= m_height ? static_cast<float>(m_width) : static_cast<float>(m_height);
 	UINT slice_count = 100; //원을 나눌 조각 수
 	float step_theta = XM_2PI / static_cast<float>(slice_count); //2PI = 360도, 조각당 각도
